psi-api.js, ?category
